refactoring of Unit of work files to match existing archetecture
