+++ Domain.Contracts/Repositories/IUnitOfWork.cs
@@ -2,5 +2,6 @@
 
 public interface IUnitOfWork
 {
+    ICourseRepository Courses { get; }
     Task CompleteAsync();
 }
