+++ LMS.Infractructure/Repositories/UnitOfWork.cs
@@ -5,11 +5,15 @@ namespace LMS.Infractructure.Repositories;
 public class UnitOfWork : IUnitOfWork
 {
     private readonly ApplicationDbContext context;
+    public ICourseRepository Courses { get; }
 
-    public UnitOfWork(ApplicationDbContext context)
+    public UnitOfWork(ApplicationDbContext context, ICourseRepository courseRepository)
     {
         this.context = context ?? throw new ArgumentNullException(nameof(context));
+        Courses = courseRepository;
     }
+
+    public async Task<int> SaveChangesAsync() => await context.SaveChangesAsync();
 
     public async Task CompleteAsync() => await context.SaveChangesAsync();
 }
